+++ templates/mentor/landing.html
@@ -19,57 +19,8 @@ Página principal - Mentor
             <h2 class="text-2xl font-bold mb-4">Solicitudes de mentoría</h2>
             {# table container #}
             {% block mentorship_req_table %}
-            <div class="border rounded-lg border-slate-400" id="table_form">
-                {# table filters #}
-                <form id="table_filters" class="flex gap-x-4 pt-4 pr-4 justify-end" hx-get="" hx-trigger="submit" hx-swap="outerHTML" hx-target="#table_form">
-                    {% include "forms/render_field.html" with field=filter_form.student_name %}
-                    {% include "forms/render_field.html" with field=filter_form.state %}
-                    <div class="self-end">
-                        {% #button type="submit" icon_left="bi bi-search" title="Filtrar y buscar" name="action" value="filter_mentorship_request" %}
-                        {% /button %}
-                    </div>
-                </form>
-                    <table id="mentorship_req_table" class="w-full">
-                        <thead>
-                            <tr>
-                                <th class="pt-6 pl-6 text-left">
-                                    Mentoría
-                                </th>
-                                <th class="pt-6 px-4 text-left">
-                                    Estudiante
-                                </th>
-                                <th class="pt-6 px-2 text-left">
-                                    Estatus
-                                </th>
-                                <th class="pt-6 pl-2 text-center">
-                                    Acción
-                                </th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {% for mentorship_request in mentorship_requests %}
-                                {% with student_user=mentorship_request.student.user %}
-                                    <tr>
-                                        <td class="pl-6 text-left">{{ mentorship_request.mentorship.name }}</td>
-                                        <td class="px-4 text-left">{{ student_user.first_name }} {{ student_user.last_name }}</td>
-                                        <td class="px-2 text-left">{{ mentorship_request.get_readable_status }}</td>
-                                        <td class="pl-2 pt-2 pb-2 flex justify-center">
-                                            {% #button type="button" icon_left="bi bi-info-circle"  %}
-                                                Revisar
-                                            {% /button %}
-                                        </td>
-                                    </tr>
-                                {% endwith %}
-                            {% empty %}
-                                <tr>
-                                    <td colspan="3" class="py-2 pl-6 text-center">No hay solicitudes de mentoría que mostrar</td>
-                                </tr>
-                            {% endfor %}
-                        </tbody>
-                    </table>
-                    <div id="pagination_container" class="flex justify-center" {# hx-get="" hx-trigger="click" hx-swap="outerHTML" hx-target="mentorship_req_table" #}>
-                        {% include "components/pagination.html" with page_obj=mentorship_requests search_query_params="action=filter_mentorship_request&"%}
-                    </div>
+            <div class="border rounded-lg border-slate-200" id="table_form">
+                {% include "mentor/mentorship_req_table.html" with with_mentorship_name="True" %}
             </div>
             {% endblock mentorship_req_table %}
 
@@ -81,6 +32,10 @@ Página principal - Mentor
             {% url 'news_feed' as news_feed_url %}
             {% include "main/event_news_feed.html" with id=NEWS_SECTION top_margin_title="True" pages=news title="Noticias" page_index=news_index load_more_url=news_feed_url page_number=news_page_number page_obj=news_paginator %}
         </section>
+    </div>
+
+    <!-- Container of the student modal -->
+    <div id="student-info-modal" data-modal-backdrop="static" tabindex="-1" aria-hidden="true" class="hidden overflow-y-auto overflow-x-hidden fixed top-0 right-0 left-0 z-50 justify-center items-center w-full md:inset-0 h-[calc(100%-1rem)] max-h-full">
     </div>
 </main>
 {% endblock body %}
