+++ templates/mentor/mentorship/request_row.html
@@ -4,18 +4,26 @@
   Row of the brief detail
   context:
     - mentorship_request: mentor.models.MentorshipRequest
+    - with_mentorship_name: bool
 {% endcomment %}
-<tr id="mentor_req_row-{{ mentorship_request.pk }}">
-    {% with student_user=mentorship_request.student.user %}
-    <td>{{ student_user.first_name }} {{ student_user.last_name }}</td>
-    {% endwith %}
-    <td class="px-4">{{ mentorship_request.get_readable_status }}</td>
-    <td>
-        <!-- Button for Opening Student Modal toggle -->
-        <div onclick="getStudentInfo('{% url 'mentor:student_info_modal' mentorship_request_pk=mentorship_request.pk %}')">
-            {% #button type="button" icon_left="bi bi-info-circle"  %}
-                Revisar
-            {% /button %}
-        </div>
-    </td>
-</tr>
+  
+{% with student_user=mentorship_request.student.user %}
+    <tr id="mentor_req_row-{{ mentorship_request.pk }}">
+        {% if with_mentorship_name %}
+            <td class="pl-6 text-left">
+                <a class="font-medium text-blue-600 hover:underline" href="{% url 'mentor:mentorship_detail' mentorship_pk=mentorship_request.mentorship.pk  %}">{{ mentorship_request.mentorship.name }}</a>
+            </td>
+        {% endif %}
+        <td class="px-4 text-left">{{ student_user.first_name }} {{ student_user.last_name }}</td>
+        <td class="px-2 text-left">{{ mentorship_request.get_readable_status }}</td>
+        <td class="pl-2 pt-2 pb-2 flex justify-center">
+            <!-- Button for Opening Student Modal toggle -->
+            <div onclick="getStudentInfo('{% url 'mentor:student_info_modal' mentorship_request_pk=mentorship_request.pk %}')">
+
+                {% #button type="button" icon_left="bi bi-info-circle"  %}
+                    Revisar
+                {% /button %}
+            </div>
+        </td>
+    </tr>
+{% endwith %}
