+++ templates/mentor/mentorship_detail.html
@@ -52,42 +52,24 @@ Mentoría: {{ mentorship.name }}
 
         <section id="solicitudes-estudiantiles" class="">
             <h2 class="text-3xl my-4">Solicitudes de mentoría</h2>
-            <table>
-                <thead>
-                    <tr>
-                        <th class="py-3 pr-4 text-left">
-                            Estudiante
-                        </th>
-                        <th class="py-3 px-4 text-left">
-                            Estatus
-                        </th>
-                        <th class="py-3 pr-4 text-left">
-                            Acción
-                        </th>
-                    </tr>
-                </thead>
-                <tbody>
-                    {% for mentorship_request in mentorship_requests %}
-                        {% include "mentor/mentorship/request_row.html" with mentorship_request=mentorship_request %}
-                    {% empty %}
-                    <tr>
-                        <td colspan="3" class="text-left">No hay solicitudes de mentoría</td>
-                    </tr>
-                    {% endfor %}
-                </tbody>
-            </table>
+
+            {% block mentorship_req_table %}
+            <div id="table_form">
+                {% include "mentor/mentorship_req_table.html" %}
+            </div>
+            {% endblock mentorship_req_table %}
         </section>
     </div>
 
     <section id="student-info">
-        <h2 class="text-2xl mt-8">Visión general de estudiantes inscritos</h2>
-        <table class="w-full border-solid border-gray-700 text-sm text-left rtl:text-right">
-            <thead class="text-xs uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
-                <tr>
-                    <th class="py-3 pr-4 text-left">
+        <h2 class="text-2xl mb-4 mt-8">Visión general de estudiantes inscritos</h2>
+        <table class="w-full border-collapse border border-ucv-yellow border-solid rounded-t-lg text-sm text-left rtl:text-right ">
+            <thead class="text-xs uppercase bg-ucv-yellow ">
+                <tr class="rounded-tl-lg rounded-tr-lg">
+                    <th class="py-3 pr-4 pl-2 text-left">
                         Nombre
                     </th>
-                    <th class="py-3 pr-4 text-left">
+                    <th class="py-3 pr-4 text-left ">
                         Estatus
                     </th>
                     <th class="py-3 pr-4 text-left"># Tareas completadas</th>
@@ -99,7 +81,7 @@ Mentoría: {{ mentorship.name }}
                 {% for student_info in students_info %}
                     <tr>
                         {# Nombre #}
-                        <td>{{ student_info.student__user__first_name }} {{ student_info.student__user__last_name }}</td>
+                        <td class="pl-2">{{ student_info.student__user__first_name }} {{ student_info.student__user__last_name }}</td>
                         {# Estatus #}
                         <td>
                             {% if student_info.is_completed %}
@@ -110,9 +92,9 @@ Mentoría: {{ mentorship.name }}
                         </td>
                         {# nro Tareas compleatadas #}
                         <td>{{ student_info.tasks_completed }}</td>
-                        <td>
+                        <td class="py-2">
                             {# TODO: link to student progress page #}
-                            <a href="{% url 'student:track_mentorship_task' mentorship_pk=mentorship.pk student_pk=student_info.student__pk %}">
+                            <a  href="{% url 'student:track_mentorship_task' mentorship_pk=mentorship.pk student_pk=student_info.student__pk %}">
                             {% #button type="button" %}
                                 Ver progreso
                             {% /button %}
@@ -122,7 +104,7 @@ Mentoría: {{ mentorship.name }}
 
                     {% empty %}
                     <tr>
-                        <td colspan="4" class="text-left">No hay estudiantes a los que les haya aprobado la mentoría</td>
+                        <td colspan="4" class="text-center pl-2 py-2">No hay estudiantes a los que les haya aprobado la mentoría</td>
                     </tr>
                 {% endfor %}
             </tbody>
